+++ SU2_CFD/include/SU2_CFD.hpp
@@ -42,7 +42,10 @@
 
 #include <ctime>
 
-#include "driver_structure.hpp"
+#include "drivers/CDriver.hpp"
+#include "drivers/CSinglezoneDriver.hpp"
+#include "drivers/CMultizoneDriver.hpp"
+#include "drivers/CDiscAdjSinglezoneDriver.hpp"
 #include "solver_structure.hpp"
 #include "integration_structure.hpp"
 #include "output_structure.hpp"
