+++ SU2_CFD/include/drivers/CDiscAdjSinglezoneDriver.hpp
@@ -1,0 +1,137 @@
+/*!
+ * \file CDiscAdjSinglezoneDriver.hpp
+ * \brief Headers of the main subroutines for driving single or multi-zone problems.
+ *        The subroutines and functions are in the <i>driver_structure.cpp</i> file.
+ * \author T. Economon, H. Kline, R. Sanchez
+ * \version 6.2.0 "Falcon"
+ *
+ * The current SU2 release has been coordinated by the
+ * SU2 International Developers Society <www.su2devsociety.org>
+ * with selected contributions from the open-source community.
+ *
+ * The main research teams contributing to the current release are:
+ *  - Prof. Juan J. Alonso's group at Stanford University.
+ *  - Prof. Piero Colonna's group at Delft University of Technology.
+ *  - Prof. Nicolas R. Gauger's group at Kaiserslautern University of Technology.
+ *  - Prof. Alberto Guardone's group at Polytechnic University of Milan.
+ *  - Prof. Rafael Palacios' group at Imperial College London.
+ *  - Prof. Vincent Terrapon's group at the University of Liege.
+ *  - Prof. Edwin van der Weide's group at the University of Twente.
+ *  - Lab. of New Concepts in Aeronautics at Tech. Institute of Aeronautics.
+ *
+ * Copyright 2012-2019, Francisco D. Palacios, Thomas D. Economon,
+ *                      Tim Albring, and the SU2 contributors.
+ *
+ * SU2 is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 2.1 of the License, or (at your option) any later version.
+ *
+ * SU2 is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with SU2. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#pragma once
+#include "CSinglezoneDriver.hpp"
+
+/*!
+ * \class CDiscAdjSinglezoneDriver
+ * \brief Class for driving single-zone adjoint solvers.
+ * \author R. Sanchez
+ * \version 6.2.0 "Falcon"
+ */
+class CDiscAdjSinglezoneDriver : public CSinglezoneDriver {
+protected:
+
+  unsigned long nAdjoint_Iter;                  /*!< \brief The number of adjoint iterations that are run on the fixed-point solver.*/
+  unsigned short RecordingState;                /*!< \brief The kind of recording the tape currently holds.*/
+  unsigned short MainVariables,                 /*!< \brief The kind of recording linked to the main variables of the problem.*/
+                 SecondaryVariables;            /*!< \brief The kind of recording linked to the secondary variables of the problem.*/
+  su2double ObjFunc;                            /*!< \brief The value of the objective function.*/
+  CIteration* direct_iteration;                 /*!< \brief A pointer to the direct iteration.*/
+
+  CConfig *config;                              /*!< \brief Definition of the particular problem. */
+  CIteration *iteration;                        /*!< \brief Container vector with all the iteration methods. */
+  CIntegration **integration;                   /*!< \brief Container vector with all the integration methods. */
+  CGeometry *geometry;                          /*!< \brief Geometrical definition of the problem. */
+  CSolver **solver;                             /*!< \brief Container vector with all the solutions. */
+
+
+public:
+
+  /*!
+   * \brief Constructor of the class.
+   * \param[in] confFile - Configuration file name.
+   * \param[in] val_nZone - Total number of zones.
+   * \param[in] val_nDim - Total number of dimensions.
+   * \param[in] MPICommunicator - MPI communicator for SU2.
+   */
+  CDiscAdjSinglezoneDriver(char* confFile,
+             unsigned short val_nZone,
+             SU2_Comm MPICommunicator);
+
+  /*!
+   * \brief Destructor of the class.
+   */
+  ~CDiscAdjSinglezoneDriver(void);
+
+  /*!
+   * \brief Preprocess the single-zone iteration
+   * \param[in] TimeIter - index of the current time-step.
+   */
+  void Preprocess(unsigned long TimeIter);
+
+  /*!
+   * \brief Run a single iteration of the discrete adjoint solver with a single zone.
+   */
+  void Run(void);
+
+  /*!
+   * \brief Postprocess the adjoint iteration for ZONE_0.
+   */
+  void Postprocess(void);
+
+  /*!
+   * \brief Record one iteration of a flow iteration in within multiple zones.
+   * \param[in] kind_recording - Type of recording (full list in ENUM_RECORDING, option_structure.hpp)
+   */
+  void SetRecording(unsigned short kind_recording);
+
+  /*!
+   * \brief Run one iteration of the solver.
+   * \param[in] kind_recording - Type of recording (full list in ENUM_RECORDING, option_structure.hpp)
+   */
+  void DirectRun(unsigned short kind_recording);
+
+  /*!
+   * \brief Set the objective function.
+   */
+  void SetObjFunction(void);
+
+  /*!
+   * \brief Initialize the adjoint value of the objective function.
+   */
+  void SetAdj_ObjFunction(void);
+
+  /*!
+   * \brief Print out the direct residuals.
+   * \param[in] kind_recording - Type of recording (full list in ENUM_RECORDING, option_structure.hpp)
+   */
+  void Print_DirectResidual(unsigned short kind_recording);
+
+  /*!
+   * \brief Record the main computational path.
+   */
+  void MainRecording(void);
+
+  /*!
+   * \brief Record the secondary computational path.
+   */
+  void SecondaryRecording(void);
+
+};
