+++ SU2_CFD/include/drivers/CDriver.hpp
@@ -38,17 +38,17 @@
 
 #pragma once
 
-#include "../../Common/include/mpi_structure.hpp"
-#include "iteration_structure.hpp"
-#include "solver_structure.hpp"
-#include "integration_structure.hpp"
-#include "output_structure.hpp"
-#include "numerics_structure.hpp"
-#include "transfer_structure.hpp"
-#include "../../Common/include/geometry_structure.hpp"
-#include "../../Common/include/grid_movement_structure.hpp"
-#include "../../Common/include/config_structure.hpp"
-#include "../../Common/include/interpolation_structure.hpp"
+#include "../../../Common/include/mpi_structure.hpp"
+#include "../iteration_structure.hpp"
+#include "../solver_structure.hpp"
+#include "../integration_structure.hpp"
+#include "../output_structure.hpp"
+#include "../numerics_structure.hpp"
+#include "../transfer_structure.hpp"
+#include "../../../Common/include/geometry_structure.hpp"
+#include "../../../Common/include/grid_movement_structure.hpp"
+#include "../../../Common/include/config_structure.hpp"
+#include "../../../Common/include/interpolation_structure.hpp"
 
 using namespace std;
 
@@ -1335,268 +1335,6 @@ public:
 
 };
 
-/*!
- * \class CSinglezoneDriver
- * \brief Class for driving single-zone solvers.
- * \author R. Sanchez
- * \version 6.0.1 "Falcon"
- */
-class CSinglezoneDriver : public CDriver {
-protected:
-
-  unsigned long TimeIter;
-
-public:
-
-  /*!
-   * \brief Constructor of the class.
-   * \param[in] confFile - Configuration file name.
-   * \param[in] val_nZone - Total number of zones.
-   * \param[in] MPICommunicator - MPI communicator for SU2.
-   */
-  CSinglezoneDriver(char* confFile,
-             unsigned short val_nZone,
-             SU2_Comm MPICommunicator);
-
-  /*!
-   * \brief Destructor of the class.
-   */
-  ~CSinglezoneDriver(void);
-
-  /*!
-   * \brief [Overload] Launch the computation for single-zone problems.
-   */
-  void StartSolver();
-
-  /*!
-   * \brief Preprocess the single-zone iteration
-   */
-  virtual void Preprocess(unsigned long TimeIter);
-
-  /*!
-   * \brief Run the iteration for ZONE_0.
-   */
-  virtual void Run();
-
-  /*!
-   * \brief Postprocess the iteration for ZONE_0.
-   */
-  virtual void Postprocess();
-
-  /*!
-   * \brief Update the dual-time solution within multiple zones.
-   */
-  void Update();
-
-  /*!
-   * \brief Output the solution in solution file.
-   */
-  void Output(unsigned long ExtIter);
-
-  /*!
-   * \brief Perform a dynamic mesh deformation, included grid velocity computation and the update of the multigrid structure.
-   */
-  void DynamicMeshUpdate(unsigned long ExtIter);
 
 
-};
-
-/*!
- * \class CDiscAdjSinglezoneDriver
- * \brief Class for driving single-zone adjoint solvers.
- * \author R. Sanchez
- * \version 6.2.0 "Falcon"
- */
-class CDiscAdjSinglezoneDriver : public CSinglezoneDriver {
-protected:
-
-  unsigned long nAdjoint_Iter;                  /*!< \brief The number of adjoint iterations that are run on the fixed-point solver.*/
-  unsigned short RecordingState;                /*!< \brief The kind of recording the tape currently holds.*/
-  unsigned short MainVariables,                 /*!< \brief The kind of recording linked to the main variables of the problem.*/
-                 SecondaryVariables;            /*!< \brief The kind of recording linked to the secondary variables of the problem.*/
-  su2double ObjFunc;                            /*!< \brief The value of the objective function.*/
-  CIteration* direct_iteration;                 /*!< \brief A pointer to the direct iteration.*/
-
-  CConfig *config;                              /*!< \brief Definition of the particular problem. */
-  CIteration *iteration;                        /*!< \brief Container vector with all the iteration methods. */
-  CIntegration **integration;                   /*!< \brief Container vector with all the integration methods. */
-  CGeometry *geometry;                          /*!< \brief Geometrical definition of the problem. */
-  CSolver **solver;                             /*!< \brief Container vector with all the solutions. */
-
-
-public:
-
-  /*!
-   * \brief Constructor of the class.
-   * \param[in] confFile - Configuration file name.
-   * \param[in] val_nZone - Total number of zones.
-   * \param[in] val_nDim - Total number of dimensions.
-   * \param[in] MPICommunicator - MPI communicator for SU2.
-   */
-  CDiscAdjSinglezoneDriver(char* confFile,
-             unsigned short val_nZone,
-             SU2_Comm MPICommunicator);
-
-  /*!
-   * \brief Destructor of the class.
-   */
-  ~CDiscAdjSinglezoneDriver(void);
-
-  /*!
-   * \brief Preprocess the single-zone iteration
-   * \param[in] TimeIter - index of the current time-step.
-   */
-  void Preprocess(unsigned long TimeIter);
-
-  /*!
-   * \brief Run a single iteration of the discrete adjoint solver with a single zone.
-   */
-  void Run(void);
-
-  /*!
-   * \brief Postprocess the adjoint iteration for ZONE_0.
-   */
-  void Postprocess(void);
-
-  /*!
-   * \brief Record one iteration of a flow iteration in within multiple zones.
-   * \param[in] kind_recording - Type of recording (full list in ENUM_RECORDING, option_structure.hpp)
-   */
-  void SetRecording(unsigned short kind_recording);
-
-  /*!
-   * \brief Run one iteration of the solver.
-   * \param[in] kind_recording - Type of recording (full list in ENUM_RECORDING, option_structure.hpp)
-   */
-  void DirectRun(unsigned short kind_recording);
-
-  /*!
-   * \brief Set the objective function.
-   */
-  void SetObjFunction(void);
-
-  /*!
-   * \brief Initialize the adjoint value of the objective function.
-   */
-  void SetAdj_ObjFunction(void);
-
-  /*!
-   * \brief Print out the direct residuals.
-   * \param[in] kind_recording - Type of recording (full list in ENUM_RECORDING, option_structure.hpp)
-   */
-  void Print_DirectResidual(unsigned short kind_recording);
-
-  /*!
-   * \brief Record the main computational path.
-   */
-  void MainRecording(void);
-
-  /*!
-   * \brief Record the secondary computational path.
-   */
-  void SecondaryRecording(void);
-
-};
-
-
-/*!
- * \class CMultizoneDriver
- * \brief Class for driving zone-specific iterations.
- * \author R. Sanchez, O. Burghardt
- * \version 6.0.1 "Falcon"
- */
-class CMultizoneDriver : public CDriver {
-protected:
-
-  bool fsi;
-  bool cht;
-
-  unsigned long TimeIter;
-
-  unsigned short *nVarZone;
-  su2double **init_res,      /*!< \brief Stores the initial residual. */
-            **residual,      /*!< \brief Stores the current residual. */
-            **residual_rel;  /*!< \brief Stores the residual relative to the initial. */
-
-  su2double flow_criteria,
-            flow_criteria_rel,
-            structure_criteria,
-            structure_criteria_rel;
-
-  bool *prefixed_motion;     /*!< \brief Determines if a fixed motion is imposed in the config file. */
-
-public:
-
-  /*!
-   * \brief Constructor of the class.
-   * \param[in] confFile - Configuration file name.
-   * \param[in] val_nZone - Total number of zones.
-   * \param[in] MPICommunicator - MPI communicator for SU2.
-   */
-  CMultizoneDriver(char* confFile,
-             unsigned short val_nZone,
-             SU2_Comm MPICommunicator);
-
-  /*!
-   * \brief Destructor of the class.
-   */
-  ~CMultizoneDriver(void);
-
-  /*!
-   * \brief [Overload] Launch the computation for multizone problems.
-   */
-  void StartSolver();
-
-  /*!
-   * \brief Preprocess the multizone iteration
-   */
-  void Preprocess(unsigned long TimeIter);
-
-  /*!
-   * \brief Use a corrector step to prevent convergence issues.
-   */
-  void Corrector(unsigned short val_iZone);
-
-  /*!
-   * \brief Run a Block Gauss-Seidel iteration in all physical zones.
-   */
-  void Run_GaussSeidel();
-
-  /*!
-   * \brief Run a Block-Jacobi iteration in all physical zones.
-   */
-  void Run_Jacobi();
-
-  /*!
-   * \brief Update the dual-time solution within multiple zones.
-   */
-  void Update();
-
-  /*!
-   * \brief Output the solution in solution file.
-   */
-  void Output(unsigned long TimeIter);
-
-  /*!
-   * \brief Check the convergence at the outer level.
-   */
-  bool OuterConvergence(unsigned long OuterIter);
-
-  /*!
-   * \brief Perform a dynamic mesh deformation, included grid velocity computation and the update of the multigrid structure (multiple zone).
-   */
-  void DynamicMeshUpdate(unsigned long ExtIter);
-
-  /*!
-   * \brief Perform a dynamic mesh deformation, including grid velocity computation and update of the multigrid structure.
-   */
-  void DynamicMeshUpdate(unsigned short val_iZone, unsigned long ExtIter);
-
-  /*!
-   * \brief Routine to provide all the desired physical transfers between the different zones during one iteration.
-   * \return Boolean that determines whether the mesh needs to be updated for this particular transfer
-   */
-  bool Transfer_Data(unsigned short donorZone, unsigned short targetZone);
-
-};
