+++ SU2_CFD/src/drivers/CDiscAdjSinglezoneDriver.cpp
@@ -35,9 +35,7 @@
  * License along with SU2. If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "../include/driver_structure.hpp"
-#include "../include/definition_structure.hpp"
-
+#include "../include/drivers/CDiscAdjSinglezoneDriver.hpp"
 
 CDiscAdjSinglezoneDriver::CDiscAdjSinglezoneDriver(char* confFile,
                                                    unsigned short val_nZone,
