+++ SU2_CFD/src/drivers/CDriver.cpp
@@ -35,7 +35,7 @@
  * License along with SU2. If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "../include/driver_structure.hpp"
+#include "../include/drivers/CDriver.hpp"
 #include "../include/definition_structure.hpp"
 
 #ifdef VTUNEPROF
